refactor: moved inline tailwind styles to external stylesheet
