+++ frontend/vite-project/src/components/credential-components/Credential.css
@@ -1,0 +1,54 @@
+.credential-form-container {
+  @apply bg-white w-full max-w-lg  min-h-96 p-8 flex flex-col;
+}
+.credential-form-header {
+  @apply mb-1;
+}
+.credential-form-subheader {
+  @apply font-medium text-lg text-[#A9ABBD] mb-4;
+}
+.credential-form {
+  @apply flex  w-full flex-col;
+}
+.credential-form-input {
+  @apply border border-black text-black  p-3 placeholder:text-[#A9ABBD] text-sm font-normal;
+}
+.email-input {
+  @apply mb-3;
+}
+.password-input {
+  @apply w-full;
+}
+.credential-eye-icon {
+  @apply absolute right-3 top-1/2 -translate-y-1/2 cursor-pointer text-gray-500;
+}
+.credential-rememberme-container {
+  @apply flex items-center gap-3;
+}
+.credential-form-checkbox {
+  @apply w-5 h-5 rounded-none cursor-pointer;
+}
+.rememberme-span-text {
+  @apply text-[#979797] font-normal text-sm leading-4 tracking-tighter;
+}
+.forgot-password-link {
+  @apply font-bold text-sm text-[#979797];
+}
+.credential-submit-btn {
+  @apply invisible h-0 w-0 p-0 m-0 border-0;
+}
+.credentials-footer {
+  @apply flex items-center justify-between;
+}
+.rememberme-label {
+  @apply flex items-center gap-3 cursor-pointer;
+}
+.social-buttons-container {
+  @apply flex flex-col mt-5 gap-3;
+}
+.social-button {
+  @apply border text-[#65676E] text-sm border-[#E1E4EB] flex justify-center items-center px-3 py-[10px] rounded-[5px];
+}
+.social-button-icon {
+  @apply text-lg mr-2 text-black;
+}
