+++ frontend/vite-project/src/pages/Signup.tsx
@@ -1,7 +1,8 @@
-import FormHeader from "../components/userAccess/FormHeader";
-import CredentialFields from "../components/userAccess/CredentialFields";
-import SocialButtons from "../components/userAccess/SocialButtons";
+import FormHeader from "../components/credential-components/FormHeader";
+import CredentialFields from "../components/credential-components/CredentialFields";
+import SocialButtons from "../components/credential-components/SocialButtons";
 import { useAuthForm } from "../hooks/useCredentialsForm";
+import "../components/credential-components/Credential.css";
 function Signup() {
   const { showPassword, handlers, formType } = useAuthForm("signup");
